+++ pnet-data-api-java/src/main/java/pnet/data/api/person/PersonDataSearch.java
@@ -21,7 +21,9 @@ import pnet.data.api.util.RestrictCompanyNumber;
 import pnet.data.api.util.RestrictCredentialsAvailable;
 import pnet.data.api.util.RestrictDatedBackUntil;
 import pnet.data.api.util.RestrictFunction;
+import pnet.data.api.util.RestrictPersonType;
 import pnet.data.api.util.RestrictQueryField;
+import pnet.data.api.util.RestrictPersonHierarchy;
 import pnet.data.api.util.RestrictRole;
 import pnet.data.api.util.RestrictTenant;
 import pnet.data.api.util.SearchWithAggregationsFunction;
@@ -33,14 +35,16 @@ import pnet.data.api.util.SearchWithAggregationsFunction;
  */
 public class PersonDataSearch
     extends AbstractSearchWithAggregations<PersonItemDTO, PersonAggregationsDTO, PersonDataSearch>
-    implements RestrictTenant<PersonDataSearch>, RestrictAdministrativeTenant<PersonDataSearch>,
-    RestrictCompanyId<PersonDataSearch>, RestrictCompanyNumber<PersonDataSearch>, RestrictCompany<PersonDataSearch>,
-    RestrictBrand<PersonDataSearch>, RestrictFunction<PersonDataSearch>, RestrictActivity<PersonDataSearch>,
-    RestrictRole<PersonDataSearch>, RestrictCredentialsAvailable<PersonDataSearch>, RestrictApproved<PersonDataSearch>,
-    RestrictDatedBackUntil<PersonDataSearch>, IncludeInactive<PersonDataSearch>, IncludeAllFunctions<PersonDataSearch>,
-    CompanyMergable<PersonDataSearch>, AggregateNumberPerTenant<PersonDataSearch>,
-    AggregateNumberPerCompany<PersonDataSearch>, AggregateNumberPerFunction<PersonDataSearch>,
-    AggregateNumberPerActivity<PersonDataSearch>, RestrictQueryField<PersonDataSearch>
+    implements RestrictTenant<PersonDataSearch>, RestrictPersonType<PersonDataSearch>,
+    RestrictAdministrativeTenant<PersonDataSearch>, RestrictCompanyId<PersonDataSearch>,
+    RestrictCompanyNumber<PersonDataSearch>, RestrictCompany<PersonDataSearch>, RestrictBrand<PersonDataSearch>,
+    RestrictFunction<PersonDataSearch>, RestrictActivity<PersonDataSearch>, RestrictRole<PersonDataSearch>,
+    RestrictPersonHierarchy<PersonDataSearch>, RestrictCredentialsAvailable<PersonDataSearch>,
+    RestrictApproved<PersonDataSearch>, RestrictDatedBackUntil<PersonDataSearch>, IncludeInactive<PersonDataSearch>,
+    IncludeAllFunctions<PersonDataSearch>, CompanyMergable<PersonDataSearch>,
+    AggregateNumberPerTenant<PersonDataSearch>, AggregateNumberPerCompany<PersonDataSearch>,
+    AggregateNumberPerFunction<PersonDataSearch>, AggregateNumberPerActivity<PersonDataSearch>,
+    RestrictQueryField<PersonDataSearch>
 {
     public PersonDataSearch(SearchWithAggregationsFunction<PersonItemDTO, PersonAggregationsDTO> searchFunction,
         List<Pair<String, Object>> restrictItems)