+++ pnet-data-api-java/src/main/java/pnet/data/api/person/PersonHierarchyType.java
@@ -1,0 +1,6 @@
+package pnet.data.api.person;
+
+public enum PersonHierarchyType
+{
+    RESPONSIBLE
+}