+++ pnet-data-api-java/src/main/java/pnet/data/api/person/PersonType.java
@@ -1,0 +1,10 @@
+package pnet.data.api.person;
+
+public enum PersonType
+{
+    REGULAR,
+
+    BOT,
+
+    TEST
+}