+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+pnet-data-api 2.8.0
+===================
+
+Needs Partner.Net 6.46.
+
+* Add person type and person hierarchy
+
 pnet-data-api 2.7.0
 ===================